Split Sphere Object
Successful split.
